+++ packages/block-library/src/navigation/edit/overlay-template-part-selector.js
@@ -299,7 +299,6 @@ export default function OverlayTemplatePartSelector( {
 						<FlexBlock>
 							<SelectControl
 								__next40pxDefaultSize
-								__nextHasNoMarginBottom
 								label={ __( 'Overlay template' ) }
 								hideLabelFromVision
 								aria-labelledby={ headingId }