+++ packages/ui/src/index.ts
@@ -5,5 +5,6 @@ export * from './form/primitives';
 export * from './icon';
 export * from './icon-button';
 export * from './stack';
+export * as Tabs from './tabs';
 export * as Tooltip from './tooltip';
 export * from './visually-hidden';