+++ packages/ui/src/utils/types.ts
@@ -1,15 +1,16 @@
-import type {
-	ElementType,
-	ComponentPropsWithoutRef,
-	HTMLAttributes,
-	ReactElement,
-	Ref,
+import {
+	type ElementType,
+	type ComponentPropsWithoutRef,
+	type HTMLAttributes,
+	type Ref,
 } from 'react';
 
 type HTMLAttributesWithRef< T extends ElementType = any > =
 	HTMLAttributes< T > & { ref?: Ref< T > | undefined };
 
-type ComponentRenderFn< Props > = ( props: Props ) => ReactElement< unknown >;
+type ComponentRenderFn< Props > = (
+	props: Props
+) => React.ReactElement< unknown >;
 
 export type ComponentProps< E extends ElementType > = Omit<
 	ComponentPropsWithoutRef< E >,
@@ -26,5 +27,5 @@ export type ComponentProps< E extends ElementType > = Omit<
 	 */
 	render?:
 		| ComponentRenderFn< HTMLAttributesWithRef >
-		| ReactElement< Record< string, unknown > >;
+		| React.ReactElement< Record< string, unknown > >;
 };